requirements.txt
Django>=4.2,<5.0
django-unfold>=0.56.0
pandas>=2.0.0
openpyxl>=3.0.0
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0
gunicorn>=21.2.0

